+++ react-app/src/components/map/clusterHelpers/Spiderfier.ts
